% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_connection}
\alias{check_connection}
\title{Checks connection}
\usage{
check_connection(connection)
}
\arguments{
\item{connection}{Connection to check. Created by \code{\link{make_connection}}}
}
\value{
Character string vector containing error messages.
}
\description{
(INTERNAL) Checks if the data given to create an inter-layer connection is valid and has the
right input format
}
\examples{
con = make_connection("mrna", "protein", connect_on="gene_name")
return_errors(check_connection(con))

}
