% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_drug_target}
\alias{check_drug_target}
\title{Check drug target interaction data}
\usage{
check_drug_target(drug_target_interaction)
}
\arguments{
\item{drug_target_interaction}{A named list of the drug interaction data. Created by
\code{\link{make_drug_target}}}
}
\value{
Character string vector containing error messages.
}
\description{
(INTERNAL) Checks if the data used to define interaction between drugs and
targets is valid and formatted correctly.
}
\examples{
data(drug_gene_interactions)
drug_target_interaction <- make_drug_target(target_molecules='protein',
interaction_table=drug_gene_interactions,
match_on='gene_name')
return_errors(check_drug_target(drug_target_interaction))
}
