% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_input}
\alias{check_input}
\title{Check pipeline input data for required format}
\usage{
check_input(layers, inter_layer_connections, drug_target_interaction)
}
\arguments{
\item{layers}{List of layers to check. Individual layers were created by
\code{\link{make_layer}} and need to be wrapped in a list.}

\item{inter_layer_connections}{A list containing connections between layers. Each
connection was created by \code{\link{make_connection}} and wrapped in a list.}

\item{drug_target_interaction}{A named list of the drug interaction data. Created by
\code{\link{make_drug_target}}}
}
\value{
Character string vector containing error messages.
}
\description{
Checks if input data is valid and formatted correctly. This function is a
wrapper for other check functions to be executed as first step of the molnet pipeline.
}
