% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_layer}
\alias{check_layer}
\title{Check layer input}
\usage{
check_layer(layer)
}
\arguments{
\item{layer}{layer to check. Created by \code{\link{make_layer}}}
}
\value{
Character string vector containing error messages.
}
\description{
(INTERNAL) Checks if the data used to create a network layer is valid and has the right
format
}
\examples{
data(metabolite_data)
metabolite_layer = make_layer(name="metabolite",
metabolite_data$group1$data,
metabolite_data$group2$data,
metabolite_data$group1$identifiers,
metabolite_data$group2$identifiers)
return_errors(check_layer(metabolite_layer))

}
