% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reduction.R
\name{chunk}
\alias{chunk}
\title{Create chunks from a vector for parallel computing}
\source{
https://stackoverflow.com/questions/3318333/split-a-vector-into-chunks
}
\usage{
chunk(x, n)
}
\arguments{
\item{x}{vector}

\item{n}{length of chunks}
}
\value{
a list of chunks of length n
}
\description{
(INTERNAL)
}
