% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_combination.R
\name{combine_graphs}
\alias{combine_graphs}
\title{Combining graphs by adding inter-layer edges}
\usage{
combine_graphs(graphs, inter_layer_edgelists)
}
\arguments{
\item{graphs}{List of iGraph objects}

\item{inter_layer_edgelists}{List of data frames containing inter-layer edges}
}
\value{
iGraph object which is the union of the input graphs with isolated nodes removed.
}
\description{
Creates the union of all graphs and adds the inter-layer edges.
}
