% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reduction.R
\name{corPvalueStudentParallel}
\alias{corPvalueStudentParallel}
\title{Compute p-values for upper triangle of correlation matrix in parallel}
\usage{
corPvalueStudentParallel(correlation, n_samples, chunk_size)
}
\arguments{
\item{correlation}{matrix of correlation values}

\item{n_samples}{matrix of number of samples used in computation of each correlation value}

\item{chunk_size}{smallest unit of work in parallel computation (number of p-values to compute)}
}
\value{
vector of p-values for upper triangle
}
\description{
(INTERNAL)
}
