% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentations.R
\docType{data}
\name{differential_score_graph_example}
\alias{differential_score_graph_example}
\title{Differential graph}
\format{
An iGraph graph object.
}
\source{
Krug, Karsten et al. “Proteogenomic Landscape of Breast Cancer Tumorigenesis and Targeted
Therapy.” Cell vol. 183,5 (2020): 1436-1456.e31. doi:10.1016/j.cell.2020.10.036
}
\usage{
differential_score_graph_example
}
\description{
Exemplary intermediate pipeline output: Differential score graph example data built by
\code{\link{differential_score}}. Contains one graph carrying the differential interaction score
as weight. This was computed using a subset of the data published by Krug et al., 2020 and
randomly sampled metabolite data \code{\link{layers_example}}.
}
\keyword{datasets}
