% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_layer}
\alias{get_layer}
\title{[INTERNAL] Fetch layer by name from layer object}
\usage{
get_layer(name, layers)
}
\arguments{
\item{name}{The layer to fetch}

\item{layers}{a layers object \code{\link{layers_example}}}
}
\value{
Returns the layer along with layer names
}
\description{
[INTERNAL] Fetch layer by name from layer object
}
