% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{get_layer_setting}
\alias{get_layer_setting}
\title{Get layer settings}
\usage{
get_layer_setting(layer, settings, setting_name)
}
\arguments{
\item{layer}{A network layer created by \code{\link{make_layer}}}

\item{settings}{Named list of settings created by \code{\link{molnet_settings}}}

\item{setting_name}{String indicating the setting to return.}
}
\value{
Setting value(s) for this layer
}
\description{
Returns specified setting for a specific network layer.
}
