% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{install_python_dependencies}
\alias{install_python_dependencies}
\title{Installs python dependencies needed for interaction score computation}
\usage{
install_python_dependencies(package_manager = "pip3")
}
\arguments{
\item{package_manager}{The package manager command or path to use (default: pip3)}
}
\description{
Uses specified pip or conda executable (default: pip3) to install all
required python modules. When using conda, the currently active
environment is used. Commands run are `pip install -r requirements` or
`conda install --file requirements`. Installs the following requirements:
- numpy
- tqdm
- python-igraph
- ray
}
