% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_combination.R
\name{inter_layer_edgelist_by_id}
\alias{inter_layer_edgelist_by_id}
\title{Interlayer conntections by identifiers}
\usage{
inter_layer_edgelist_by_id(annotation_1, annotation_2, connection, weight = 1)
}
\arguments{
\item{annotation_1, annotation_2}{Data frames: Annotation tables specifying the identifiers of
the nodes of a network}

\item{connection}{String of identifier to connect on}

\item{weight}{Integer or vector specifying the weight of the inter-layer connections.}
}
\value{
data.frame with columns from, to and weight
}
\description{
(INTERNAL) Returns an edge list defining the connections between two layers of the network.
}
