% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentations.R
\docType{data}
\name{interaction_score_graphs_example}
\alias{interaction_score_graphs_example}
\title{Interaction score graphs}
\format{
A named list with 2 items.
\describe{
  \item{group1}{iGraph graph object containing the interaction score as weight for group1.}
  \item{group2}{}
}
}
\source{
Krug, Karsten et al. “Proteogenomic Landscape of Breast Cancer Tumorigenesis and Targeted
 Therapy.” Cell vol. 183,5 (2020): 1436-1456.e31. doi:10.1016/j.cell.2020.10.036
}
\usage{
interaction_score_graphs_example
}
\description{
Exemplary intermediate pipeline output: Interaction score graphs example data built by
\code{\link{interaction_score}}. A named list (elements `group1` and `group2`). Each element
contains an iGraph object containing the interaction score as weight. This was computed using a
subset of the data published by Krug et al., 2020 and randomly sampled metabolite data
\code{\link{layers_example}}.
}
\keyword{datasets}
