% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentations.R
\docType{data}
\name{interaction_score_graphs_vignette}
\alias{interaction_score_graphs_vignette}
\title{Interaction score graphs for vignette}
\format{
A named list with 2 items.
\describe{
  \item{group1}{iGraph graph object containing the interaction score as weight for group1.}
  \item{group2}{}
}
}
\source{
Krug, Karsten et al. “Proteogenomic Landscape of Breast Cancer Tumorigenesis and Targeted
Therapy.” Cell vol. 183,5 (2020): 1436-1456.e31. doi:10.1016/j.cell.2020.10.036
}
\usage{
interaction_score_graphs_vignette
}
\description{
Exemplary intermediate pipeline output used in the vignette.
}
\keyword{datasets}
