% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_score.R
\name{load_interaction_score_output}
\alias{load_interaction_score_output}
\title{Loads output of python script for interaction score calculation}
\usage{
load_interaction_score_output(loading_path)
}
\arguments{
\item{loading_path}{Directory to load from}
}
\value{
A named list (elements `group1` and `group2`). Each element contains an iGraph object
containing the interaction score as edge attribute.
}
\description{
(INTERNAL) Loads data generated by \code{\link{calculate_interaction_score}}. Output files are
graphs in gml format for both groups.
}
