% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_pickHardThreshold.R
\name{scaleFreeFitIndex_alternative}
\alias{scaleFreeFitIndex_alternative}
\title{Alternative implementation of WGCNA::scaleFreeFitIndex}
\source{
\code{\link[WGCNA]{scaleFreeFitIndex}}
}
\usage{
scaleFreeFitIndex_alternative(k, nBreaks = 10, removeFirst = FALSE)
}
\arguments{
\item{k}{numeric vector whose components contain non-negative values}

\item{nBreaks}{positive integer. This determines the number of equal width bins.}

\item{removeFirst}{logical. If TRUE then the first bin will be removed.}
}
\value{
the model fitting index (R.squared) from the following model lm(log.p.dk ~ log.dk)
}
\description{
(INTERNAL) This function is a copy of \code{\link[WGCNA]{scaleFreeFitIndex}}
with minor changes introduced to fit the needs of this package.

Description by \code{\link[WGCNA]{scaleFreeFitIndex}}: The function scaleFreeFitIndex
calculates several indices (fitting statistics) for evaluating scale free topology fit. The
input is a vector (of connectivities) k. Next k is discretized into nBreaks number of equal-
width bins. Let's denote the resulting vector dk. The relative frequency for each bin is
denoted p.dk.

The entire original function code is contained in the function and deleted lines are
commented out using \code{#--} to show changes.
}
