% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_score.R
\name{write_interaction_score_input}
\alias{write_interaction_score_input}
\title{Write edge lists and combined graphs to files}
\usage{
write_interaction_score_input(
  combined_graphs,
  drug_target_edgelists,
  saving_path
)
}
\arguments{
\item{combined_graphs}{A named list (elements `group1` and `group2`). Each element contains the
entire combined network (layers + inter-layer connections) as iGraph graph object.}

\item{drug_target_edgelists}{A named list (elements `group1` and `group2`). Each element
contains the list of edges to be considered in the interaction score calculation as data frame (columns `from`, `to` and `weight`)}

\item{saving_path}{Directory to write to}
}
\value{
Does not return value, but writes to .tsv.
}
\description{
(INTERNAL) Writes the combined graphs and the drug target edge lists to files for
passing them to the python interaction score script.
Graphs are saved as `gml` file. Edgelists are saved as `tsv` file.
}
